[Refactor] #643 - App.vue Sidebar를 common/SideBar.vue로 분리
- Sidebar 템플릿, 메뉴 데이터, 네비게이션 로직을 SideBar.vue로 이동
- App.vue 312줄 → 102줄로 축소
- components/common 디렉토리 신규 생성
